pyproject.toml
[project]
name = "reddit-video-maker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "chatterbox-tts>=0.1.2",
    "forcealign>=1.1.9",
    "moviepy>=2.2.1",
    "num2words>=0.5.14",
    "ollama>=0.5.1",
    "praw>=7.8.1",
]
